requirements.txt
torch
torchvision
torchaudio
hydra-core
tensorboard
tqdm
scipy
matplotlib